changed TRGCDC::getEventTime to integer type
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@28830 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ trg/cdc/include/EventTime.h
@@ -47,7 +47,7 @@ namespace Belle2{
       void hist(void);
       void oldVer(void);
       void printFirm(void);
-      double getT0(void)const;
+      int getT0(void)const;
 
    
     private :
